Finished Main Menu (Mostly)
The basic Main Menu has been mostly implemented. I don't know if we want to actually add a save/load feature, but I have the buttons set as if we are. I don't think that I will be able to get to finish the save/load if we want it.